geolocation-API/getCurrentPosition_permission_deny.https.html

<!DOCTYPE HTML>
<meta charset='utf-8'>
<title>Geolocation Test: getCurrentPosition location access denied</title>
<link rel='help' href='http://www.w3.org/TR/geolocation-API/#privacy_for_uas'>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src='/resources/testdriver.js'></script>
<script src='/resources/testdriver-vendor.js'></script>
<script src='support.js'></script>
<script>
// Rewrite http://dev.w3.org/geo/api/test-suite/t.html?00001
promise_test(async () => {
  await test_driver.set_permission({name: 'geolocation'}, 'denied');

  let hasMethodReturned = false;

  return new Promise((resolve, reject) => {
    geo.getCurrentPosition(resolve, reject);
    hasMethodReturned = true;
  }).then(pos => {
    assert_unreached('A success callback was invoked unexpectedly with position ' + positionToString(pos));
  }).catch(err => {
    // http://dev.w3.org/geo/api/test-suite/t.html?00031
    assert_true(hasMethodReturned, 'Check that getCurrentPosition returns synchronously before any callbacks are invoked');
    assert_equals(err.code, err.PERMISSION_DENIED, errorToString(err));
  });
}, 'User denies access, check that error callback is called with correct code');
</script>
